+++ PostHog/PostHogConfig.swift
@@ -170,8 +170,9 @@ public typealias BeforeSendBlock = (PostHogEvent) -> PostHogEvent?
         @objc public let sessionReplayConfig: PostHogSessionReplayConfig = .init()
     #endif
 
-    /// Configuration for error tracking (Experimental)
-    @_spi(Experimental)
+    /// Configuration for error tracking.
+    ///
+    /// See known limitations: https://posthog.com/docs/error-tracking/installation/ios#limitations
     @objc public let errorTrackingConfig: PostHogErrorTrackingConfig = .init()
 
     /// Enable mobile surveys